+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 - Add Azure RangeReaders [#386](https://github.com/geotrellis/geotrellis-server/pull/386)
 - Add GetFeatureInfoExtended support [#388](https://github.com/geotrellis/geotrellis-server/pull/388)
+- Add Rasterio Affine Transform support [#390](https://github.com/geotrellis/geotrellis-server/pull/390)
 
 ## Changed
 - STAC Assets regex selectors support [#388](https://github.com/geotrellis/geotrellis-server/pull/388)